requirements.txt
streamlit==1.32.0
altair==4.2.2
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
pillow==10.3.0
tensorflow==2.12.0
keras==2.12.0
joblib==1.3.2